requirements.txt
biopython>=1.79
click>=8.0
pandas>=1.3
matplotlib>=3.4
seaborn>=0.11
pyyaml>=5.4
tqdm>=4.65.0
numpy>=1.21.0